+++ frontend/src/App.css
@@ -1053,6 +1053,7 @@ input.hidden {
   right: 10px;
   width: 820px;
   border: 1px solid #d9d9d9;
+  border-radius: 7px;
 }
 
 .filterbox:before {
