+++ frontend/src/components/Filter/FilterByVisitType.js
