+++ frontend/src/components/Filter/FilterByVisitType.spec.js
